Fixed some minor bugs, did some reorganizing!
