+++ src/main/java/frc/util/pid/SparkMaxConstants.java
@@ -1,4 +1,4 @@
-package frc.util.constants;
+package frc.util.pid;
 
 public class SparkMaxConstants {
     //#region turret
